Improve identifying keys as character keys
The `.key != .code` approach leads to wrongly capturing some unusual
special keys as well as numpad keys used in non-numeric mode, and
opening a search bar with the key name filled in. Non-character key
names seem to consist of alphanumeric ASCII characters, so checking for
that might be a more reliable heuristic.